requirements.txt
cffi==1.17.1
charset-normalizer==3.4.0
cryptography==43.0.1
pdfminer.six==20231228
pdfplumber==0.11.4
pillow==10.4.0
pycparser==2.22
pypdfium2==4.30.0
python-dotenv==1.0.1
